voice-cloning-pipeline/requirements.txt
# Voice Cloning Pipeline Requirements

# Core Azure SDKs
azure-cognitiveservices-speech>=1.38.0
azure-storage-blob>=12.22.0

# AI & ML
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# Web Framework (for dashboard)
flask>=3.0.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.11.0
librosa>=0.10.0
soundfile>=0.12.0
pandas>=2.0.0
matplotlib>=3.7.0
pydub>=0.25.1
